+++ libnavui-voice/src/main/java/com/mapbox/navigation/ui/voice/api/MapboxAudioGuidance.kt
@@ -43,6 +43,7 @@ internal constructor(
 
     private var dataStoreOwner: NavigationDataStoreOwner? = null
     private var configOwner: NavigationConfigOwner? = null
+    private var audioGuidanceVoice: MapboxAudioGuidanceVoice? = null
     private var mutedStateFlow = MutableStateFlow(false)
     private val internalStateFlow = MutableStateFlow(MapboxAudioGuidanceState())
     private val scope = CoroutineScope(SupervisorJob() + dispatcher)
@@ -74,6 +75,7 @@ internal constructor(
      */
     override fun onDetached(mapboxNavigation: MapboxNavigation) {
         mapboxVoiceInstructions.unregisterObservers(mapboxNavigation)
+        audioGuidanceVoice?.destroy()
         job?.cancel()
         job = null
     }
@@ -164,15 +166,24 @@ internal constructor(
         }
     }
 
-    private fun MapboxNavigation.audioGuidanceVoice(): Flow<MapboxAudioGuidanceVoice> =
-        combine(
+    private fun MapboxNavigation.audioGuidanceVoice(): Flow<MapboxAudioGuidanceVoice> {
+        var trigger: VoiceInstructionsDownloadTrigger? = null
+        return combine(
             mapboxVoiceInstructions.voiceLanguage(),
             configOwner!!.language(),
         ) { voiceLanguage, deviceLanguage -> voiceLanguage ?: deviceLanguage }
             .distinctUntilChanged()
             .map { language ->
-                audioGuidanceServices.mapboxAudioGuidanceVoice(this, language)
+                audioGuidanceVoice?.destroy()
+                trigger?.let { unregisterVoiceInstructionsTriggerObserver(it) }
+                audioGuidanceServices.mapboxAudioGuidanceVoice(this, language).also {
+                    audioGuidanceVoice = it
+                    trigger = VoiceInstructionsDownloadTrigger(it.mapboxSpeechApi).also {
+                        registerVoiceInstructionsTriggerObserver(it)
+                    }
+                }
             }
+    }
 
     private suspend fun restoreMutedState() {
         dataStoreOwner?.apply {