+++ libnavui-voice/src/main/java/com/mapbox/navigation/ui/voice/api/TimeBasedNextVoiceInstructionsProvider.kt
@@ -1,7 +1,5 @@
 package com.mapbox.navigation.ui.voice.api
 
-import com.mapbox.api.directions.v5.models.LegStep
-import com.mapbox.api.directions.v5.models.RouteLeg
 import com.mapbox.api.directions.v5.models.VoiceInstructions
 import com.mapbox.navigation.utils.internal.logW
 
@@ -27,22 +25,25 @@ internal class TimeBasedNextVoiceInstructionsProvider(
             }
 
         val voiceInstructions = mutableListOf<VoiceInstructions>()
-        fillCurrentStepVoiceInstructions(
-            currentStep,
-            progress.stepDistanceRemaining,
-            voiceInstructions
-        )
         var cumulatedTime = progress.stepDurationRemaining
+        val currentStepInstructions = currentStep.voiceInstructions()?.filter { instruction ->
+            val distanceAlongGeometry = instruction.distanceAlongGeometry()
+            distanceAlongGeometry != null &&
+                distanceAlongGeometry <= progress.stepDistanceRemaining
+        }
+        if (currentStepInstructions != null) {
+            voiceInstructions.addAll(currentStepInstructions)
+        }
 
-        // fill next steps
         var currentStepIndex = progress.stepIndex
         var currentLegIndex = progress.legIndex
         while (cumulatedTime < observableTimeSeconds) {
-            if (isLastStep(currentStepIndex, legSteps)) {
+            if (currentStepIndex + 1 < (legSteps?.size ?: 0)) {
+                currentStep = legSteps!![currentStepIndex + 1]
+                currentStepIndex++
+            } else {
                 currentStepIndex = 0
-                if (isLastLeg(currentLegIndex, legs)) {
-                    break
-                } else {
+                if (currentLegIndex + 1 < legs.size) {
                     legSteps = legs[currentLegIndex + 1].steps()
                     currentLegIndex++
                     if (legSteps.isNullOrEmpty()) {
@@ -50,38 +51,14 @@ internal class TimeBasedNextVoiceInstructionsProvider(
                     } else {
                         currentStep = legSteps.first()
                     }
+                } else {
+                    break
                 }
-            } else {
-                currentStep = legSteps!![currentStepIndex + 1]
-                currentStepIndex++
             }
             currentStep.voiceInstructions()?.let { voiceInstructions.addAll(it) }
             cumulatedTime += currentStep.duration()
         }
         return voiceInstructions
-    }
-
-    private fun fillCurrentStepVoiceInstructions(
-        currentStep: LegStep,
-        stepDistanceRemaining: Double,
-        voiceInstructions: MutableList<VoiceInstructions>
-    ) {
-        val currentStepInstructions = currentStep.voiceInstructions()?.filter { instruction ->
-            val distanceAlongGeometry = instruction.distanceAlongGeometry()
-            distanceAlongGeometry != null &&
-                distanceAlongGeometry <= stepDistanceRemaining
-        }
-        if (currentStepInstructions != null) {
-            voiceInstructions.addAll(currentStepInstructions)
-        }
-    }
-
-    private fun isLastStep(currentStepIndex: Int, legSteps: List<LegStep>?): Boolean {
-        return currentStepIndex + 1 >= (legSteps?.size ?: 0)
-    }
-
-    private fun isLastLeg(currentLegIndex: Int, legs: List<RouteLeg>): Boolean {
-        return currentLegIndex + 1 >= legs.size
     }
 
     private companion object {