+++ libnavui-voice/src/test/java/com/mapbox/navigation/ui/voice/api/MapboxVoiceApiTest.kt
@@ -1,8 +1,8 @@
 package com.mapbox.navigation.ui.voice.api
 
+import com.mapbox.api.directions.v5.models.VoiceInstructions
 import com.mapbox.bindgen.ExpectedFactory
 import com.mapbox.navigation.ui.voice.model.SpeechAnnouncement
-import com.mapbox.navigation.ui.voice.model.TypeAndAnnouncement
 import com.mapbox.navigation.ui.voice.model.VoiceState
 import com.mapbox.navigation.ui.voice.testutils.Fixtures
 import io.mockk.coEvery
@@ -22,26 +22,25 @@ import java.io.InputStream
 internal class MapboxVoiceApiTest {
 
     private lateinit var sut: MapboxVoiceApi
-    private lateinit var mockSpeechProvider: MapboxSpeechProvider
+    private lateinit var mockSpeechLoder: MapboxSpeechLoader
     private lateinit var mockSpeechFileProvider: MapboxSpeechFileProvider
 
     @Before
     fun setUp() {
-        mockSpeechProvider = mockk(relaxed = true)
+        mockSpeechLoder = mockk(relaxed = true)
         mockSpeechFileProvider = mockk(relaxed = true)
 
-        sut = MapboxVoiceApi(mockSpeechProvider, mockSpeechFileProvider)
+        sut = MapboxVoiceApi(mockSpeechLoder, mockSpeechFileProvider)
     }
 
     @Test
     fun `retrieveVoiceFile should download audio data using MapboxSpeechProvider`() = runBlocking {
         val voiceInstructions = Fixtures.ssmlInstructions()
-        coEvery { mockSpeechProvider.load(any()) } returns ExpectedFactory.createError(Error())
+        coEvery { mockSpeechLoder.load(any(), any()) } returns ExpectedFactory.createError(Error())
 
-        sut.retrieveVoiceFile(voiceInstructions)
+        sut.retrieveVoiceFile(voiceInstructions, true)
 
-        val announcement = TypeAndAnnouncement("ssml", voiceInstructions.ssmlAnnouncement()!!)
-        coVerify { mockSpeechProvider.load(announcement) }
+        coVerify { mockSpeechLoder.load(voiceInstructions, true) }
     }
 
     @Test
@@ -50,12 +49,12 @@ internal class MapboxVoiceApiTest {
             val voiceInstructions = Fixtures.ssmlInstructions()
             val blob = byteArrayOf(11, 22)
             val blobInputStream = slot<InputStream>()
-            coEvery { mockSpeechProvider.load(any()) } returns ExpectedFactory.createValue(blob)
+            coEvery { mockSpeechLoder.load(any(), false) } returns ExpectedFactory.createValue(blob)
             coEvery {
                 mockSpeechFileProvider.generateVoiceFileFrom(capture(blobInputStream))
             } returns File("ignored")
 
-            sut.retrieveVoiceFile(voiceInstructions)
+            sut.retrieveVoiceFile(voiceInstructions, false)
 
             assertEquals(blob.count(), blobInputStream.captured.available())
         }
@@ -66,10 +65,10 @@ internal class MapboxVoiceApiTest {
             val voiceInstructions = Fixtures.ssmlInstructions()
             val blob = byteArrayOf(11, 22)
             val file = File("saved-audio-file")
-            coEvery { mockSpeechProvider.load(any()) } returns ExpectedFactory.createValue(blob)
+            coEvery { mockSpeechLoder.load(any(), true) } returns ExpectedFactory.createValue(blob)
             coEvery { mockSpeechFileProvider.generateVoiceFileFrom(any()) } returns file
 
-            val result = sut.retrieveVoiceFile(voiceInstructions)
+            val result = sut.retrieveVoiceFile(voiceInstructions, true)
 
             assertEquals(VoiceState.VoiceFile(file), result)
         }
@@ -78,10 +77,12 @@ internal class MapboxVoiceApiTest {
     fun `retrieveVoiceFile should return VoiceError on any error`() =
         runBlocking {
             val voiceInstructions = Fixtures.emptyInstructions()
-            coEvery { mockSpeechProvider.load(any()) } returns ExpectedFactory.createError(Error())
+            coEvery {
+                mockSpeechLoder.load(any(), false)
+            } returns ExpectedFactory.createError(Error())
             coEvery { mockSpeechFileProvider.generateVoiceFileFrom(any()) } throws Error()
 
-            val result = sut.retrieveVoiceFile(voiceInstructions)
+            val result = sut.retrieveVoiceFile(voiceInstructions, false)
 
             assertTrue(result is VoiceState.VoiceError)
         }
@@ -92,12 +93,26 @@ internal class MapboxVoiceApiTest {
         val mockedFile: File = mockk()
         every { mockedAnnouncement.file } returns mockedFile
         val fileProvider: MapboxSpeechFileProvider = mockk(relaxed = true)
-        val speechProvider: MapboxSpeechProvider = mockk()
-        val mapboxVoiceApi = MapboxVoiceApi(speechProvider, fileProvider)
+        val speechLoader: MapboxSpeechLoader = mockk()
+        val mapboxVoiceApi = MapboxVoiceApi(speechLoader, fileProvider)
 
         mapboxVoiceApi.clean(mockedAnnouncement)
 
         verify(exactly = 1) { fileProvider.delete(mockedFile) }
+    }
+
+    @Test
+    fun predownload() {
+        val fileProvider = mockk<MapboxSpeechFileProvider>(relaxed = true)
+        val speechLoader = mockk<MapboxSpeechLoader>(relaxed = true)
+        val mapboxVoiceApi = MapboxVoiceApi(speechLoader, fileProvider)
+        val instructions = listOf(mockk<VoiceInstructions>())
+
+        mapboxVoiceApi.predownload(instructions)
+
+        verify(exactly = 1) {
+            speechLoader.triggerDownload(instructions)
+        }
     }
 
     @Test
@@ -106,8 +121,8 @@ internal class MapboxVoiceApiTest {
         val nullFile: File? = null
         every { mockedAnnouncement.file } returns nullFile
         val fileProvider: MapboxSpeechFileProvider = mockk(relaxed = true)
-        val speechProvider: MapboxSpeechProvider = mockk()
-        val mapboxVoiceApi = MapboxVoiceApi(speechProvider, fileProvider)
+        val speechLoader: MapboxSpeechLoader = mockk()
+        val mapboxVoiceApi = MapboxVoiceApi(speechLoader, fileProvider)
 
         mapboxVoiceApi.clean(mockedAnnouncement)
 
@@ -117,11 +132,18 @@ internal class MapboxVoiceApiTest {
     @Test
     fun cancel() {
         val fileProvider = mockk<MapboxSpeechFileProvider>(relaxed = true)
-        val speechProvider = mockk<MapboxSpeechProvider>()
-        val mapboxVoiceApi = MapboxVoiceApi(speechProvider, fileProvider)
+        val speechLoader = mockk<MapboxSpeechLoader>()
+        val mapboxVoiceApi = MapboxVoiceApi(speechLoader, fileProvider)
 
         mapboxVoiceApi.cancel()
 
         verify(exactly = 1) { fileProvider.cancel() }
+    }
+
+    @Test
+    fun `destroy cancels speech loader`() {
+        sut.destroy()
+
+        verify { mockSpeechLoder.cancel() }
     }
 }